requirements.txt
streamlit>=1.32.0
anthropic>=0.40.0
groq>=0.11.0
google-genai>=0.3.0
duckdb>=1.0.0
pandas>=2.0.0
